fix(guard): break shim gate import cycle
