+++ src/codex_plugin_scanner/guard/shim_probe.py
@@ -3,35 +3,10 @@
 from __future__ import annotations
 
 import json
-import shlex
-from collections.abc import Sequence
-from pathlib import Path
 from typing import Any
 
 SHIM_PROBE_ENV_VAR = "HOL_GUARD_SHIM_PROBE"
 SHIM_PROBE_ENV_VALUE = "1"
-_PACKAGE_SHIM_PARSER_MANAGERS = frozenset(
-    {
-        "bun",
-        "bundle",
-        "cargo",
-        "composer",
-        "go",
-        "gradle",
-        "mvn",
-        "npm",
-        "npx",
-        "pip",
-        "pip3",
-        "pipenv",
-        "pipx",
-        "pnpm",
-        "poetry",
-        "uv",
-        "uvx",
-        "yarn",
-    }
-)
 
 _PACKAGE_SHIM_PROBE_ARGS: dict[str, tuple[str, ...]] = {
     "npm": ("install", "--dry-run", "lodash@4.17.21"),
@@ -56,26 +31,6 @@ def package_shim_probe_args(manager: str) -> tuple[str, ...]:
     """Return install-shaped probe args that route through package protect."""
 
     return _PACKAGE_SHIM_PROBE_ARGS.get(manager, ("--version",))
-
-
-def package_shim_command_requires_guard(
-    manager: str,
-    argv: Sequence[str],
-    *,
-    workspace: Path | None = None,
-) -> bool:
-    """Return whether a shimmed package-manager command should enter Guard protect."""
-
-    normalized_manager = manager.strip().lower()
-    if not normalized_manager:
-        return False
-    if normalized_manager not in _PACKAGE_SHIM_PARSER_MANAGERS:
-        return True
-    command = [normalized_manager, *[str(argument) for argument in argv]]
-    from codex_plugin_scanner.guard.runtime.package_intent_parser import parse_package_intent
-
-    intent = parse_package_intent(shlex.join(command), workspace=workspace)
-    return intent is not None
 
 
 def parse_protect_json_stdout(stdout: str) -> dict[str, Any]:
